diet-recipe-planner/backend/requirements.txt
# FastAPI web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# HTTP client for LLM APIs
httpx==0.28.1
openai==1.59.6

# Telegram bot
python-telegram-bot==21.9
python-telegram-bot[job-queue]==21.9

# Scheduler
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
